Build parse queue and stats GUI editors. Add skeleton for Spectrum Exporter.
Created two display elements that update upon BIN file selection: one displays all the files in the data series that are to be parsed, the other (will eventually) display the run statistics for the full series (such as the real/live times, etc). Also added a new Group tab in the GUI that is to be the Spectrum Exporter. Currently, just a vanilla file browser sits waiting.