Add SpectrumExporter class to model; migrate write_spec functionality
Incomplete class that's not ready for running. Cleaning up
functionality that's been migrated from write_spec, now a method for
the Exporter class. Attention needs to be paid to how the output stings
are constructed and writtten to the file. The 'times' dictionary has
also undergone changes that need to be addressed in class methods.